Added SaleTest + misc changes.
Amount and VAT changed to BigDecimal data type and changes to make the conversion work.
Getters added to SaleSummaryDTO.

+++ iv1350/src/se.kth.iv1350.DTO/ItemDTO.java
@@ -3,6 +3,8 @@ package se.kth.iv1350.DTO;
 import se.kth.iv1350.model.Amount;
 import se.kth.iv1350.model.VAT;
 
+import java.math.BigDecimal;
+
 public class ItemDTO {
     private Amount price;
     private VAT vatRate;
@@ -24,7 +26,7 @@ public class ItemDTO {
         return price;
     }
 
-    public float getVatRate() {
+    public BigDecimal getVatRate() {
         return vatRate.getRate();
     }
